==> Step_2_ACWR_PCWR.ipynb <==
This part of the code estimated the Peak Scheme Water Demand (PSWD) and the Average/Seasonal Scheme Water Demand (SSWD) based on crop water requirements that have been computer in Step 1


# import the required Python packages
import pandas as pd
import datetime   # check this
import numpy as np
import ast       # check this
from ast import literal_eval  # check this
from pandas import DataFrame

# from pyeto import fao
from datetime import datetime
%matplotlib inline     # check this

data=pd.read_excel('Pilot_Result_Part1.xlsx') #This is the output file from code1. 
#data

#To add a new coloumn for date palms irrigated area (1ha=10000m2  and *assuming that date palm area = 50% of total area)

data['harv_i_ha']= data['area_m2']/(10000)

%%time

#Setting the default value for these parameters


for i in range (1,13):
    data['ACWR_{}'.format(i)]=0       #ACWR: Average Crop Water Requirement in (m3/ha), Previously CWN.
    data['PCWR_{}'.format(i)]=0       #PCWR: Peak Crop Water Requirement (l/s/ha) or "Duty", Previously PDWR
    data['harvested_{}'.format(i)]=0  #This repsents the actual area harvested in each month depending on the growing season.
    data['PWD_{}'.format(i)]=0        #PWD: Peak Water Demand in (l/s)
    data['SSWD_{}'.format(i)]=0       #SSWD: Seasonal Scheme Water Demand in (m3)

%%time
#STEP 1: Compute the ACWR from ETc - check FAO1992- page 43-

#acwr=row['ETo_{}'.format(i)]*30*row['kc_{}'.format(i)] - row['eff_{}'.format(i)]*30 (( - row['awc']/12 ))
#once the available water content layer is obtained, the last past should be added to the equation

for i in range(1, 13):
    eto = f'ETo_{i}'
    kc = f'kc_{i}'
    eff = f'eff_{i}'
    acwr = f'ACWR_{i}'
    data[acwr] = data[eto]*30*data[kc] - data[eff]*30
    data.loc[data[acwr]<0,acwr] = 0
        

%%time
# STEP 2: Computing the PCWR (l/s/ha) 

# The following notes expatins the equaitons: 
# 1. Crop Water Requirment, CWR (m3/ha) = CWR (mm) * 10 
# 2. Average Crop Water Requirement (m3/ha/d) = (CWR (m3/ha) / Length of crop duration (days))
# 3. Assuming here that the lenght of the season is 30 (1 month)
# 4. Assuming that the Peak Crop Water Requirement = 2 * average crop water requirement (m3/ha/d)
# 5. conversion factor from (m3/d/ha) to (l/sec/ha) is 0.012
# Naming changed from (crop water need,CWN) in the initial code to (crop water requirement) to be consistence with the source FAO1992. 



for i in range(1,13):
    data['PCWR_{}'.format(i)]=((data['ACWR_{}'.format(i)]*10)/30)*2*0.012

%%time

for index,row in data.iterrows():
    len_init1= (len(range(row['init1_start_month'],row['init1_end_month']))+1)
    len_init2= (len(range(row['init2_start_month'],row['init2_end_month']))+1)
    len_init= (len_init1)+(len_init2)
#len_init2= len(range(row['init2_start_month'],row['init2_end_month']))+1
#len_init=len_init1 + len_init2
#len_init1=(row['init1_start_month'] - row['init2_end_month'])
#print(len_init)


#
#
#
#
#

#%%time

#STEPs 3 and 4: Estimating Peak Water Demand (PWD) in (l/s) and Seasonal Scheme Water Demand (SSWD) in (m3) 
# In order to estimate PWD and SSWS we need first to compute the irrigated area used in that particular month/season. 

#PWD = PCWR / Irrigation efficiency(IrrEff) ; IrrEff = Field Application Efficiency (aeff) * Distribution Efficiency (deff)*100 
#deff (distribution efficieny %): 0.95 (all scenarios)
#aeff (field application efficiency %) : 0.6 (SU), 0.75 (SP), 0.9 (DR)

##Scenario option might be: Surface irrigation aeff = 0.6, Sprinkler aeff = 0.75 and Drip aeff = 09. 

count_p=0 #To adjust the count of months in the loop below
count_h=0 #To adjust the count of months in the loop below
pumping_hours_per_day=10 #is this an assumption?? 
deff= 1
aeff= 0.6

init1_count = np.zeros(len(data))
init2_count = np.zeros(len(data))
late_count = np.zeros(len(data))

for i in [11,12,1,2,3,4,5,6,7,8,9,10]:

    init1 = [(i >= j) & (i <= k) for j, k in zip(data['init1_start_month'],data['init1_end_month'])]
    init2 = [(i >= j) & (i <= k) for j, k in zip(data['init2_start_month'],data['init2_end_month'])]

    init1_count += init1 * 1
    init2_count += init2 * 1
    init_count = init1_count + init2_count

    data.loc[np.array(init1) | np.array(init2),'harvested_{}'.format(i)] =(data['harv_i_ha']/(len_init)*init_count)
    data.loc[np.array(init1) | np.array(init2),'PWD_{}'.format(i)]= (data['PCWR_{}'.format(i)] *(data['harvested_{}'.format(i)]*24))/(pumping_hours_per_day*aeff*deff)
    data.loc[np.array(init1) | np.array(init2),'SSWD_{}'.format(i)]= (data['ACWR_{}'.format(i)]*10*(data['harvested_{}'.format(i)])/(aeff*deff))
    
    dev = [(i >= j) & (i <= k) for j, k in zip(data['dev_start_month'],data['dev_end_month'])]
    
    data.loc[dev,'harvested_{}'.format(i)]=data['harv_i_ha']
    data.loc[dev,'PWD_{}'.format(i)]=(data['PCWR_{}'.format(i)]*data['harv_i_ha']*24)/(pumping_hours_per_day*aeff*deff)
    data.loc[dev,'SSWD_{}'.format(i)]= (data['ACWR_{}'.format(i)]*10*data['harv_i_ha'])/(aeff*deff)
    
    
    mid = [(i >= j) & (i <= k) for j, k in zip(data['mid_start_month'],data['mid_end_month'])]
    
    data.loc[mid,'harvested_{}'.format(i)]=data['harv_i_ha']
    data.loc[mid,'PWD_{}'.format(i)]=(data['PCWR_{}'.format(i)]*data['harv_i_ha']*24)/(pumping_hours_per_day*aeff*deff)
    data.loc[mid,'SSWD_{}'.format(i)]= (data['ACWR_{}'.format(i)]*10*data['harv_i_ha'])/(aeff*deff)
    
    late = [(i >= j) & (i <= k) for j, k in zip(data['late_start_month'],data['late_end_month'])]

    late_count += late * 1

    data.loc[late,'harvested_{}'.format(i)]=(data['harv_i_ha']/([len(range(i,j+1)) for i,j in zip(data['late_start_month'],data['late_end_month'])])*late_count)
    data.loc[late,'PWD_{}'.format(i)]= (data['PCWR_{}'.format(i)]*(data['harvested_{}'.format(i)]*24)/(pumping_hours_per_day*aeff*deff))
    data.loc[late,'SSWD_{}'.format(i)]= (data['ACWR_{}'.format(i)]*10*(data['harvested_{}'.format(i)])/(aeff*deff))


#Create a Pandas Excel writer using XlsxWriter as the engine.
writer = pd.ExcelWriter('Pilot_Result_Part2.xlsx', engine='xlsxwriter')

# Convert the dataframe to an XlsxWriter Excel object.
data.to_excel(writer, sheet_name='test_all')


# Close the Pandas Excel writer and output the Excel file.
writer.save()
